# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsignmentPayablesReport": {
        "description": "寄售應付報表（docs/19 §2.5）。\n\n只計 PENDING 入待付合計；PAID/CANCELLED 分欄；reclaim_needed（已付後退貨需追回）獨立分欄，\n不以負數沖抵 pending。status_filter 只影響明細列，合計恆涵蓋全部狀態。",
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ConsignmentPayableRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "status_filter": {
            "title": "Status Filter",
            "type": "string"
          },
          "store_id": {
            "title": "Store Id",
            "type": "integer"
          },
          "total_cancelled_payout": {
            "title": "Total Cancelled Payout",
            "type": "string"
          },
          "total_paid_payout": {
            "title": "Total Paid Payout",
            "type": "string"
          },
          "total_pending_payout": {
            "title": "Total Pending Payout",
            "type": "string"
          },
          "total_reclaim_needed_payout": {
            "title": "Total Reclaim Needed Payout",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "store_id",
          "status_filter",
          "rows",
          "total_pending_payout",
          "total_paid_payout",
          "total_cancelled_payout",
          "total_reclaim_needed_payout"
        ],
        "title": "ConsignmentPayablesReport",
        "type": "object"
      },
...
      "DailyCashReport": {
        "description": "每日現金對帳報表（docs/19 §2.2）。expected 與關帳 expected_amount 同源。",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/DailyCashSessionRow"
            },
            "title": "Sessions",
            "type": "array"
          },
          "store_id": {
            "title": "Store Id",
            "type": "integer"
          },
          "total_acquisition_void_in": {
            "title": "Total Acquisition Void In",
            "type": "string"
          },
          "total_buyout_out": {
            "title": "Total Buyout Out",
            "type": "string"
          },
          "total_cash_sales": {
            "title": "Total Cash Sales",
            "type": "string"
          },
          "total_consignment_payout_out": {
            "title": "Total Consignment Payout Out",
            "type": "string"
          },
          "total_counted": {
            "title": "Total Counted",
            "type": "string"
          },
          "total_expected": {
            "title": "Total Expected",
            "type": "string"
          },
          "total_manual_adjust": {
            "title": "Total Manual Adjust",
            "type": "string"
          },
          "total_opening_float": {
            "title": "Total Opening Float",
            "type": "string"
          },
          "total_sale_refund_out": {
            "title": "Total Sale Refund Out",
            "type": "string"
          },
          "total_store_credit_redeemed_display_only": {
            "title": "Total Store Credit Redeemed Display Only",
            "type": "string"
          },
          "total_variance": {
            "title": "Total Variance",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "store_id",
          "date",
          "sessions",
          "total_opening_float",
          "total_cash_sales",
          "total_acquisition_void_in",
          "total_buyout_out",
          "total_consignment_payout_out",
          "total_sale_refund_out",
          "total_manual_adjust",
          "total_expected",
          "total_counted",
          "total_variance",
          "total_store_credit_redeemed_display_only"
        ],
        "title": "DailyCashReport",
        "type": "object"
      },
...
      "DailySummaryReport": {
        "description": "每日營運儀表板（docs/19 R5）：組合 R1 現金 + R2 毛利的同源數字，店長一眼看「今天賺多少」。\n\n營業額 vs 認列營收明確區分（寄售全額 ≠ 店家營收）；成本未知不假造毛利；估算淨利明確標註。",
        "properties": {
          "acquisition_void_in": {
            "title": "Acquisition Void In",
            "type": "string"
          },
          "avg_ticket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ticket"
          },
          "buyout_out": {
            "title": "Buyout Out",
            "type": "string"
          },
          "cash_sales_in": {
            "title": "Cash Sales In",
            "type": "string"
          },
          "cash_variance": {
            "title": "Cash Variance",
            "type": "string"
          },
          "cogs": {
            "title": "Cogs",
            "type": "string"
          },
          "consignment_commission_income": {
            "title": "Consignment Commission Income",
            "type": "string"
          },
          "consignment_payout_out": {
            "title": "Consignment Payout Out",
            "type": "string"
          },
          "counted_cash": {
            "title": "Counted Cash",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "estimated_net_income": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Net Income"
          },
          "estimated_net_income_note": {
            "title": "Estimated Net Income Note",
            "type": "string"
          },
          "expected_cash": {
            "title": "Expected Cash",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "gross_margin": {
            "title": "Gross Margin",
            "type": "string"
          },
          "gross_margin_rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Margin Rate"
          },
          "gross_turnover": {
            "title": "Gross Turnover",
            "type": "string"
          },
          "manual_adjust": {
            "title": "Manual Adjust",
            "type": "string"
          },
          "net_sales_ex_tax": {
            "title": "Net Sales Ex Tax",
            "type": "string"
          },
          "recognized_revenue": {
            "title": "Recognized Revenue",
            "type": "string"
          },
          "store_credit_issued": {
            "title": "Store Credit Issued",
            "type": "string"
          },
          "store_credit_redeemed": {
            "title": "Store Credit Redeemed",
            "type": "string"
          },
          "store_id": {
            "title": "Store Id",
            "type": "integer"
          },
          "tax": {
            "title": "Tax",
            "type": "string"
          },
          "total_cash_out": {
            "title": "Total Cash Out",
            "type": "string"
          },
          "transaction_count": {
            "title": "Transaction Count",
            "type": "integer"
          },
          "unknown_cost_sales": {
            "title": "Unknown Cost Sales",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "store_id",
          "date",
          "gross_turnover",
          "recognized_revenue",
          "net_sales_ex_tax",
          "tax",
          "consignment_commission_income",
          "cogs",
          "gross_margin",
          "gross_margin_rate",
          "unknown_cost_sales",
          "cash_sales_in",
          "acquisition_void_in",
          "buyout_out",
          "consignment_payout_out",
          "manual_adjust",
          "total_cash_out",
          "expected_cash",
          "counted_cash",
          "cash_variance",
          "store_credit_issued",
          "store_credit_redeemed",
          "transaction_count",
          "avg_ticket",
          "estimated_net_income",
          "estimated_net_income_note"
        ],
        "title": "DailySummaryReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TrendsReport": {
        "description": "財務趨勢時間序列（docs/19 R6）：依 granularity 分桶的 R5 同義 KPI；餵趨勢圖。\n\n桶與 [from, to) 取交集（首/末桶可為部分期間），故各桶 KPI 加總 = 全期 margin_breakdown，\n可交叉驗證（同源）。空桶補 0 列，圖表連續。日界一律 UTC。",
        "properties": {
          "date_from": {
            "format": "date-time",
            "title": "Date From",
            "type": "string"
          },
          "date_to": {
            "format": "date-time",
            "title": "Date To",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "granularity": {
            "title": "Granularity",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TrendRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "store_id": {
            "title": "Store Id",
            "type": "integer"
          }
        },
        "required": [
          "generated_at",
          "store_id",
          "date_from",
          "date_to",
          "granularity",
          "rows"
        ],
        "title": "TrendsReport",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "lu-camp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reports/consignment-payables": {
      "get": {
        "description": "寄售應付（docs/19 §2.5）：只計 PENDING 待付；PAID/CANCELLED/reclaim 分欄；不輸出身分證。",
        "operationId": "consignmentPayablesReport",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "ALL",
              "enum": [
                "PENDING",
                "PAID",
                "CANCELLED",
                "ALL"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "xlsx"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsignmentPayablesReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consignment Payables",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/daily-cash": {
      "get": {
        "description": "每日現金對帳（docs/19 §2.2）：依 session 分列 + 當日合計；expected 與關帳同公式。",
        "operationId": "dailyCashReport",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "xlsx"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCashReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Daily Cash",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/daily-summary": {
      "get": {
        "description": "每日營運儀表板（docs/19 R5）：今日營業額/認列營收/毛利/現金支出/購物金/估算淨利一覽。",
        "operationId": "dailySummaryReport",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "xlsx"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySummaryReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Daily Summary",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/inventory-value": {
      "get": {
        "description": "庫存價值與庫齡（docs/19 §2.4）：自有成本/售價、寄售在庫另列、catalog 成本 N/A、自有庫齡。",
        "operationId": "inventoryValueReport",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "xlsx"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryValueReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Inventory Value",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/sales-margin": {
      "get": {
        "description": "銷售 / 毛利（docs/19 §2.3）。半開區間 [from, to)；to<=from → 422。",
        "operationId": "salesMarginReport",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "xlsx"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesMarginReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sales Margin",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/v1/reports/trends": {
      "get": {
        "description": "財務趨勢時間序列（docs/19 R6）：daily/weekly/monthly/quarterly KPI，餵趨勢圖。半開區間。",
        "operationId": "financeTrendsReport",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "month",
              "enum": [
                "day",
                "week",
                "month",
                "quarter"
              ],
              "title": "Granularity",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "xlsx"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trends",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}